+++ packages/bluesky/lib/src/entities/converter/thread_rule_converter.dart
@@ -2,9 +2,11 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+// 📦 Package imports:
 import 'package:atproto_core/atproto_core.dart' as core;
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// 🌎 Project imports:
 import '../keys/ids.g.dart' as ids;
 import '../thread_mention_rule.dart';
 import '../thread_rule.dart';