+++ packages/bluesky/lib/src/entities/thread_list_rule.g.dart
@@ -14,16 +14,17 @@ _$_ThreadListRule _$$_ThreadListRuleFromJson(Map json) => $checkedCreate(
       ($checkedConvert) {
         final val = _$_ThreadListRule(
           type: $checkedConvert(
-              'type', (v) => v as String? ?? appBskyFeedThreadgateListRule),
-          list: $checkedConvert(
+              r'$type', (v) => v as String? ?? appBskyFeedThreadgateListRule),
+          listUri: $checkedConvert(
               'list', (v) => atUriConverter.fromJson(v as String)),
         );
         return val;
       },
+      fieldKeyMap: const {'type': r'$type', 'listUri': 'list'},
     );
 
 Map<String, dynamic> _$$_ThreadListRuleToJson(_$_ThreadListRule instance) =>
     <String, dynamic>{
-      'type': instance.type,
-      'list': atUriConverter.toJson(instance.list),
+      r'$type': instance.type,
+      'list': atUriConverter.toJson(instance.listUri),
     };